Финальная версия верстки 4 страниц(Главная, Логин, Регистрация, NotFound) Сделать Code Review.
