+++ src/pages/NotFound/NotFound.jsx
@@ -1,6 +1,5 @@
 import * as S from './NotFound.styles.js';
 import { useNavigate } from 'react-router-dom';
-//import { ListOfCourses } from '../ListOfCourses/ListOfCourses.jsx';
  
 export const NotFoundPage = () => {
 
@@ -10,13 +9,13 @@ export const NotFoundPage = () => {
     }
   
     return (
-      <>
+    <S.Conteiner>
       <S.TitelNotfound>
         <S.TitelNotfoundErrorText>404</S.TitelNotfoundErrorText>
         <S.TitelNotfoundErrorInformation>Страница не найдена<img src="img/smilecrying.png" alt="smilecrying" /></S.TitelNotfoundErrorInformation>
         <S.TitelNotfoundInfo>Возможно, она была удалена<br/> или перенесена на другой адрес</S.TitelNotfoundInfo>
         <S.NotfoundButton onClick={onClick}>Вернуться на главную</S.NotfoundButton>
       </S.TitelNotfound>
-    </>
+    </S.Conteiner>
     )
   }
