+++ src/routes.jsx
@@ -4,13 +4,10 @@ import { SignUp } from './pages/signUp/SignUp.jsx';
 import { ListOfCourses } from './pages/ListOfCourses/ListOfCourses.jsx';
 import { NotFoundPage } from './pages/NotFound/NotFound.jsx';
 
-export const AppRoutes = ({ setIsLoading, isLoading }) => {
+export const AppRoutes = () => {
     return (
     <Routes>
-        <Route path='/' element={<ListOfCourses 
-                                   isLoading={isLoading} 
-                                   setIsLoading={setIsLoading}
-                                />}></Route>
+        <Route path='/' element={<ListOfCourses />}></Route>
         <Route path="/login" element={<Login />}></Route>
         <Route path="/SignUp" element={<SignUp />}></Route>
         <Route path="*" element={<NotFoundPage />}></Route>
